Avoid deadlock in the schedule function
A deadlock can occur when a timer expires, and a user thread that owns
a spinlock is scheduled simultaneously. To avoid this deadlock, the
scheduling of the native thread is disabled when a user thread owns
a spinlock.

The thread-local 'preempt_disable_count' variable is used to determine
if the thread can be preempted or not.

Patched by <colinyoyo26@gmail.com>